Move function only used for buffer, prune another method
